Added ItemPriceUtils and moved all item price functions into it.
